fix: Clean up logger transport ownership
I think there may be a bug in the way winston-transport's typescript
type definitions are written. This makes modules that depend on this
module freak out when they have to directly deal with a transport. As
long as we don't own any transports, things seem to work. This patch
allows a transport to take references to members of our logging
classes, instead of the logging class owning the transport AND the
logger.