chore: Split core into core and core-backend
Due to new security restrictions in electron, things get very unhappy
when node types creep into the render process. The core library was
a mix of frontend and backend code, but needs to be separated so that
we don't get any pollution between the processes, otherwise things
won't even start.

Core is now the frontend library, and core-backend is all the required
node glue.